@@ -1,33 +1,33 @@
 {
   "name": "thingsboard-js-executor",
   "private": true,
-  "version": "3.1.0",
+  "version": "3.1.1",
   "description": "ThingsBoard JavaScript Executor Microservice",
   "main": "server.js",
   "bin": "server.js",
   "scripts": {
-    "install": "pkg -t node10-linux-x64,node10-win-x64 --out-path ./target . && node install.js",
+    "install": "pkg -t node12-linux-x64,node12-win-x64 --out-path ./target . && node install.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon server.js",
     "start-prod": "NODE_ENV=production nodemon server.js"
   },
   "dependencies": {
-    "@azure/service-bus": "^1.1.7",
-    "@google-cloud/pubsub": "^1.7.3",
-    "amqplib": "^0.5.6",
-    "aws-sdk": "^2.677.0",
+    "@azure/service-bus": "^1.1.9",
+    "@google-cloud/pubsub": "^2.5.0",
+    "amqplib": "^0.6.0",
+    "aws-sdk": "^2.741.0",
     "azure-sb": "^0.11.1",
     "config": "^3.3.1",
-    "js-yaml": "^3.12.0",
+    "js-yaml": "^3.14.0",
     "kafkajs": "^1.12.0",
     "long": "^4.0.0",
-    "uuid-parse": "^1.0.0",
-    "uuid-random": "^1.3.0",
-    "winston": "^3.0.0",
-    "winston-daily-rotate-file": "^3.2.1"
+    "uuid-parse": "^1.1.0",
+    "uuid-random": "^1.3.2",
+    "winston": "^3.3.3",
+    "winston-daily-rotate-file": "^4.5.0"
   },
   "engines": {
-    "node": ">=8.0.0 <11.0.0"
+    "node": ">=12.0.0 <14.0.0"
   },
   "nyc": {
     "exclude": [
@@ -38,9 +38,9 @@
     ]
   },
   "devDependencies": {
-    "fs-extra": "^6.0.1",
-    "nodemon": "^1.17.5",
-    "pkg": "^4.4.8"
+    "fs-extra": "^9.0.1",
+    "nodemon": "^2.0.4",
+    "pkg": "^4.4.9"
   },
   "pkg": {
     "assets": [
