@@ -1,29 +1,29 @@
 {
   "name": "thingsboard-web-ui",
   "private": true,
-  "version": "3.1.0",
+  "version": "3.1.1",
   "description": "ThingsBoard Web UI Microservice",
   "main": "server.js",
   "bin": "server.js",
   "scripts": {
-    "install": "pkg -t node10-linux-x64,node10-win-x64 --out-path ./target . && node install.js",
+    "install": "pkg -t node12-linux-x64,node12-win-x64 --out-path ./target . && node install.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "WEB_FOLDER=./target/web nodemon server.js",
     "start-prod": "NODE_ENV=production nodemon server.js"
   },
   "dependencies": {
-    "compression": "^1.7.3",
-    "config": "^3.2.2",
-    "connect-history-api-fallback": "^1.5.0",
-    "express": "^4.16.3",
+    "compression": "^1.7.4",
+    "config": "^3.3.1",
+    "connect-history-api-fallback": "^1.6.0",
+    "express": "^4.17.1",
     "http": "0.0.0",
-    "http-proxy": "^1.17.0",
-    "js-yaml": "^3.12.0",
-    "winston": "^3.0.0",
-    "winston-daily-rotate-file": "^3.2.1"
+    "http-proxy": "^1.18.1",
+    "js-yaml": "^3.14.0",
+    "winston": "^3.3.3",
+    "winston-daily-rotate-file": "^4.5.0"
   },
   "engines": {
-    "node": ">=8.0.0 <11.0.0"
+    "node": ">=12.0.0 <14.0.0"
   },
   "nyc": {
     "exclude": [
@@ -34,9 +34,9 @@
     ]
   },
   "devDependencies": {
-    "fs-extra": "^6.0.1",
-    "nodemon": "^1.17.5",
-    "pkg": "^4.4.0"
+    "fs-extra": "^9.0.1",
+    "nodemon": "^2.0.4",
+    "pkg": "^4.4.9"
   },
   "pkg": {
     "assets": [
