@@ -13,123 +12,124 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.7",
-    "@angular/cdk": "^9.2.4",
-    "@angular/common": "^9.1.7",
-    "@angular/compiler": "^9.1.7",
-    "@angular/core": "^9.1.7",
-    "@angular/flex-layout": "^9.0.0-beta.31",
-    "@angular/forms": "^9.1.7",
-    "@angular/material": "^9.2.4",
-    "@angular/platform-browser": "^9.1.7",
-    "@angular/platform-browser-dynamic": "^9.1.7",
-    "@angular/router": "^9.1.7",
-    "@auth0/angular-jwt": "^4.0.0",
+    "@angular/animations": "^10.0.9",
+    "@angular/cdk": "^10.1.3",
+    "@angular/common": "^10.0.9",
+    "@angular/compiler": "^10.0.9",
+    "@angular/core": "^10.0.9",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "^10.0.9",
+    "@angular/material": "^10.1.3",
+    "@angular/platform-browser": "^10.0.9",
+    "@angular/platform-browser-dynamic": "^10.0.9",
+    "@angular/router": "^10.0.9",
+    "@auth0/angular-jwt": "^5.0.1",
     "@date-io/date-fns": "^2.6.1",
-    "@flowjs/flow.js": "^2.14.0",
-    "@flowjs/ngx-flow": "^0.4.3",
+    "@flowjs/flow.js": "^2.14.1",
+    "@flowjs/ngx-flow": "^0.4.4",
     "@juggle/resize-observer": "^3.1.3",
-    "@mat-datetimepicker/core": "^4.1.0",
-    "@material-ui/core": "^4.9.13",
+    "@mat-datetimepicker/core": "^5.0.1",
+    "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/pickers": "^3.2.10",
-    "@ngrx/effects": "^9.1.2",
-    "@ngrx/store": "^9.1.2",
-    "@ngrx/store-devtools": "^9.1.2",
-    "@ngx-share/core": "^7.1.4",
-    "@ngx-translate/core": "^12.1.2",
-    "@ngx-translate/http-loader": "^4.0.0",
-    "ace-builds": "^1.4.11",
-    "angular-gridster2": "^9.1.0",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@ngrx/store-devtools": "^10.0.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/http-loader": "^6.0.0",
+    "ace-builds": "^1.4.12",
+    "angular-gridster2": "^10.1.3",
     "angular2-hotkeys": "^2.2.0",
-    "base64-js": "^1.3.1",
     "canvas-gauges": "^2.1.7",
     "compass-sass-mixins": "^0.12.7",
     "core-js": "^3.6.5",
-    "date-fns": "^2.12.0",
+    "date-fns": "^2.15.0",
     "flot": "git://github.com/thingsboard/flot.git#0.9-work",
     "flot.curvedlines": "git://github.com/MichaelZinsmaier/CurvedLines.git#master",
     "font-awesome": "^4.7.0",
     "jquery": "^3.5.1",
     "jquery.terminal": "^2.16.0",
-    "js-beautify": "^1.11.0",
+    "js-beautify": "^1.12.0",
     "json-schema-defaults": "^0.4.0",
-    "jstree": "^3.3.9",
+    "jstree": "^3.3.10",
     "jstree-bootstrap-theme": "^1.0.1",
-    "jszip": "^3.4.0",
+    "jszip": "^3.5.0",
     "leaflet": "^1.6.0",
     "leaflet-editable": "^1.2.0",
     "leaflet-polylinedecorator": "^1.6.0",
-    "leaflet-providers": "^1.10.1",
+    "leaflet-providers": "^1.10.2",
     "leaflet.gridlayer.googlemutant": "0.10.0",
     "leaflet.markercluster": "^1.4.1",
     "material-design-icons": "^3.0.1",
     "messageformat": "^2.3.0",
-    "moment": "^2.24.0",
+    "moment": "^2.27.0",
     "ngx-clipboard": "^13.0.1",
-    "ngx-color-picker": "^9.1.0",
-    "ngx-daterangepicker-material": "^3.0.1",
-    "ngx-flowchart": "git://github.com/thingsboard/ngx-flowchart.git#e3b694b73f44d3170ee1f9ca7f3de6f6ec92cc71",
+    "ngx-color-picker": "^10.0.1",
+    "ngx-daterangepicker-material": "^3.0.4",
+    "ngx-flowchart": "git://github.com/thingsboard/ngx-flowchart.git#master",
     "ngx-hm-carousel": "^2.0.0-rc.1",
-    "ngx-translate-messageformat-compiler": "^4.6.0",
+    "ngx-sharebuttons": "^8.0.1",
+    "ngx-translate-messageformat-compiler": "^4.8.0",
     "objectpath": "^2.0.0",
+    "prettier": "^2.0.5",
     "prop-types": "^15.7.2",
     "raphael": "^2.3.0",
-    "rc-select": "^10.2.5",
+    "rc-select": "^11.1.3",
     "react": "^16.13.1",
-    "react-ace": "^8.1.0",
+    "react-ace": "^9.1.3",
     "react-dom": "^16.13.1",
-    "react-dropzone": "^10.2.2",
+    "react-dropzone": "^11.0.3",
     "reactcss": "^1.2.3",
-    "rxjs": "^6.5.5",
-    "schema-inspector": "1.6.8",
+    "rxjs": "^6.6.2",
+    "schema-inspector": "^1.7.0",
     "screenfull": "^5.0.2",
-    "split.js": "^1.5.11",
+    "split.js": "^1.6.2",
     "systemjs": "0.21.5",
     "tinycolor2": "^1.4.1",
-    "tooltipster": "^4.2.7",
-    "tslib": "^1.11.1",
+    "tooltipster": "^4.2.8",
+    "tslib": "^2.0.1",
     "tv4": "^1.3.0",
     "typeface-roboto": "^0.0.75",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^9.1.0",
-    "@angular-devkit/build-angular": "^0.901.6",
-    "@angular/cli": "^9.1.6",
-    "@angular/compiler-cli": "^9.1.7",
-    "@angular/language-service": "^9.1.7",
+    "@angular-builders/custom-webpack": "^10.0.0",
+    "@angular-devkit/build-angular": "^0.1000.5",
+    "@angular/cli": "^10.0.5",
+    "@angular/compiler-cli": "^10.0.9",
+    "@angular/language-service": "^10.0.9",
     "@types/canvas-gauges": "^2.1.2",
     "@types/flot": "^0.0.31",
-    "@types/jasmine": "^3.5.10",
+    "@types/jasmine": "^3.5.12",
     "@types/jasminewd2": "^2.0.8",
-    "@types/jquery": "^3.3.38",
-    "@types/js-beautify": "^1.8.2",
+    "@types/jquery": "^3.5.1",
+    "@types/js-beautify": "^1.11.0",
     "@types/jstree": "^3.3.40",
     "@types/jszip": "^3.4.1",
     "@types/leaflet": "^1.5.17",
     "@types/leaflet-markercluster": "^1.0.3",
     "@types/leaflet-polylinedecorator": "^1.6.0",
-    "@types/lodash": "^4.14.151",
+    "@types/lodash": "^4.14.159",
     "@types/raphael": "^2.3.0",
-    "@types/react": "^16.9.35",
-    "@types/react-dom": "^16.9.7",
+    "@types/react": "^16.9.46",
+    "@types/react-dom": "^16.9.8",
     "@types/tinycolor2": "^1.4.2",
-    "@types/tooltipster": "^0.0.29",
-    "codelyzer": "^5.2.2",
-    "compression-webpack-plugin": "^3.1.0",
+    "@types/tooltipster": "^0.0.30",
+    "codelyzer": "^6.0.0",
+    "compression-webpack-plugin": "^4.0.1",
     "directory-tree": "^2.2.4",
-    "jasmine-core": "^3.5.0",
-    "jasmine-spec-reporter": "^5.0.1",
-    "karma": "^5.0.2",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^2.1.1",
-    "karma-jasmine": "^3.1.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "~5.1.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
     "ngrx-store-freeze": "^0.2.4",
-    "protractor": "^5.4.4",
-    "ts-node": "^8.9.0",
-    "tslint": "^6.1.1",
-    "typescript": "~3.7.5"
+    "protractor": "~7.0.0",
+    "ts-node": "^8.10.2",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.7",
+    "webpack": "^4.44.1"
   }
 }
