@@ -25,6 +25,8 @@ import { DashboardUtilsService } from '@core/services/dashboard-utils.service';
 import { DashboardResolver } from '@app/modules/home/pages/dashboard/dashboard-routing.module';
 import { UtilsService } from '@core/services/utils.service';
 import { Widget } from '@app/shared/models/widget.models';
+import { MODULES_MAP } from '../../shared/models/constants';
+import { modulesMap } from '../common/modules-map';
 
 @Injectable()
 export class WidgetEditorDashboardResolver implements Resolve<Dashboard> {
@@ -92,7 +94,11 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
   providers: [
-    WidgetEditorDashboardResolver
+    WidgetEditorDashboardResolver,
+    {
+      provide: MODULES_MAP,
+      useValue: modulesMap
+    }
   ]
 })
 export class DashboardPagesRoutingModule { }
