@@ -197,7 +197,7 @@ export class EntitiesTableWidgetComponent extends PageComponent implements OnIni
     if (this.displayPagination) {
       this.sort.sortChange.subscribe(() => this.paginator.pageIndex = 0);
     }
-    (this.displayPagination ? merge(this.sort.sortChange, this.paginator.page) : this.sort.sortChange)
+    ((this.displayPagination ? merge(this.sort.sortChange, this.paginator.page) : this.sort.sortChange) as Observable<any>)
       .pipe(
         tap(() => this.updateData())
       )
