@@ -50,7 +50,7 @@ export type MapSettings = {
     markerClusteringSetting?;
     useDefaultCenterPosition?: boolean;
     gmDefaultMapType?: string;
-    useLabelFunction: string;
+    useLabelFunction: boolean;
     icon?: any;
     zoomOnClick: boolean,
     maxZoom: number,
@@ -62,7 +62,7 @@ export type MapSettings = {
     useCustomProvider: boolean,
     customProviderTileUrl: string;
     mapPageSize: number;
-}
+};
 
 export enum MapProviders {
     google = 'google-map',
@@ -103,7 +103,7 @@ export type MarkerSettings = {
     markerImageFunction?: MarkerImageFunction;
     markerOffsetX: number;
     markerOffsetY: number;
-}
+};
 
 export interface FormattedData {
     $datasource: Datasource;
@@ -112,19 +112,19 @@ export interface FormattedData {
     entityType: EntityType;
     dsIndex: number;
     deviceType: string;
-    [key: string]: any
+    [key: string]: any;
 }
 
 export interface ReplaceInfo {
   variable: string;
   valDec?: number;
-  dataKeyName: string
+  dataKeyName: string;
 }
 
 export type PolygonSettings = {
     showPolygon: boolean;
     polygonKeyName: string;
-    polKeyName: string;// deprecated
+    polKeyName: string; // deprecated
     polygonStrokeOpacity: number;
     polygonOpacity: number;
     polygonStrokeWeight: number;
@@ -141,7 +141,7 @@ export type PolygonSettings = {
     polygonTooltipFunction: GenericFunction;
     polygonColorFunction?: GenericFunction;
     editablePolygon: boolean;
-}
+};
 
 export type PolylineSettings = {
     usePolylineDecorator: any;
@@ -166,7 +166,7 @@ export type PolylineSettings = {
     colorFunction: GenericFunction;
     strokeOpacityFunction: GenericFunction;
     strokeWeightFunction: GenericFunction;
-}
+};
 
 export interface HistorySelectSettings {
     buttonColor: string;
@@ -196,7 +196,7 @@ export type TripAnimationSettings = {
     pointAsAnchorFunction: GenericFunction;
     tooltipFunction: GenericFunction;
     labelFunction: GenericFunction;
-}
+};
 
 export type actionsHandler = ($event: Event, datasource: Datasource) => void;
 
@@ -244,4 +244,5 @@ export const hereProviders = [
     'HERE.normalDay',
     'HERE.normalNight',
     'HERE.hybridDay',
-    'HERE.terrainDay']
+    'HERE.terrainDay'
+];
