@@ -57,7 +57,7 @@ export class Polyline {
           })
         }
       ]
-    }
+    };
   }
 
   updatePolyline(locations: L.LatLng[], data: FormattedData, dataSources: FormattedData[], settings: PolylineSettings) {
@@ -65,8 +65,9 @@ export class Polyline {
     this.dataSources = dataSources;
     this.leafletPoly.setLatLngs(locations);
     this.leafletPoly.setStyle(this.getPolyStyle(settings));
-    if (this.polylineDecorator)
+    if (this.polylineDecorator) {
       this.polylineDecorator.setPaths(this.leafletPoly);
+    }
   }
 
   getPolyStyle(settings: PolylineSettings): L.PolylineOptions {
@@ -78,7 +79,7 @@ export class Polyline {
         [this.data, this.dataSources, this.data.dsIndex], settings.strokeOpacity),
       weight: functionValueCalculator(settings.useStrokeWeightFunction, settings.strokeWeightFunction,
         [this.data, this.dataSources, this.data.dsIndex], settings.strokeWeight)
-    }
+    };
   }
 
   removePolyline() {
