@@ -23,7 +23,7 @@ import { StateControllerComponent } from './state-controller.component';
 import { StatesControllerService } from '@home/pages/dashboard/states/states-controller.service';
 import { EntityId } from '@app/shared/models/id/entity-id';
 import { UtilsService } from '@core/services/utils.service';
-import { base64toObj, insertVariable, isEmpty, objToBase64 } from '@app/core/utils';
+import { base64toObj, insertVariable, isEmpty, objToBase64URI } from '@app/core/utils';
 import { DashboardUtilsService } from '@core/services/dashboard-utils.service';
 import { EntityService } from '@core/http/entity.service';
 import { EntityType } from '@shared/models/entity-type.models';
@@ -281,7 +281,7 @@ export class EntityStateControllerComponent extends StateControllerComponent imp
       if (this.isDefaultState()) {
         newState = null;
       } else {
-        newState = objToBase64(this.stateObject);
+        newState = objToBase64URI(this.stateObject);
       }
       this.updateStateParam(newState);
     }
