@@ -14,7 +14,7 @@
 /// limitations under the License.
 ///
 
-import { OnDestroy } from '@angular/core';
+import { Directive, OnDestroy } from '@angular/core';
 import { PageComponent } from '@shared/components/page.component';
 import { Store } from '@ngrx/store';
 import { AppState } from '@core/core.state';
@@ -23,6 +23,7 @@ import { NavigationStart, Router, RouterEvent } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
 
+@Directive()
 export abstract class DialogComponent<T, R = any> extends PageComponent implements OnDestroy {
 
   routerSubscription: Subscription;
