@@ -14,19 +14,12 @@
 /// limitations under the License.
 ///
 
-import { JsonSettingsSchema } from '@shared/models/widget.models';
-import { MapProviders } from './map-models';
-
-export interface MapWidgetInterface {
-    resize();
-    update();
-    onInit();
-    onDestroy();
-}
-
-export interface MapWidgetStaticInterface {
-    settingsSchema(mapProvider?: MapProviders, drawRoutes?: boolean): JsonSettingsSchema;
-    getProvidersSchema(mapProvider?: MapProviders, ignoreImageMap?: boolean): JsonSettingsSchema;
-    dataKeySettingsSchema(): object;
-    actionSources(): object;
+export function enumerable(value: boolean) {
+  return (
+    target: any,
+    propertyKey: string,
+    descriptor: PropertyDescriptor
+  ) => {
+    descriptor.enumerable = value;
+  };
 }
